+++ herbst/db/schema/room.go
@@ -1,0 +1,30 @@
+package schema
+
+import (
+	"entgo.io/ent"
+	"entgo.io/ent/schema/edge"
+	"entgo.io/ent/schema/field"
+)
+
+// Room holds the schema definition for the Room entity.
+type Room struct {
+	ent.Schema
+}
+
+// Fields of the Room.
+func (Room) Fields() []ent.Field {
+	return []ent.Field{
+		field.String("name"),
+		field.String("description"),
+		field.Bool("isStartingRoom").
+			Default(false),
+		field.JSON("exits", map[string]int{}),
+	}
+}
+
+// Edges of the Room.
+func (Room) Edges() []ent.Edge {
+	return []ent.Edge{
+		edge.To("characters", Character.Type),
+	}
+}